Extracted common part of the project in commons module

+++ order-service/src/main/java/com/example/orderservice/dto/response/CreateOrderItemResponseDTO.java
@@ -1,10 +1,12 @@
 package com.example.orderservice.dto.response;
 
+import java.math.BigDecimal;
+
 public class CreateOrderItemResponseDTO {
 
     private Long productId;
     private Integer quantity;
-    private Double priceAtPurchase;
+    private BigDecimal priceAtPurchase;
 
     public CreateOrderItemResponseDTO() {
     }
@@ -25,11 +27,11 @@ public class CreateOrderItemResponseDTO {
         this.quantity = quantity;
     }
 
-    public Double getPriceAtPurchase() {
+    public BigDecimal getPriceAtPurchase() {
         return priceAtPurchase;
     }
 
-    public void setPriceAtPurchase(Double priceAtPurchase) {
+    public void setPriceAtPurchase(BigDecimal priceAtPurchase) {
         this.priceAtPurchase = priceAtPurchase;
     }
 }
