+++ order-service/src/main/java/com/example/orderservice/entity/OrderItem.java
@@ -2,6 +2,8 @@ package com.example.orderservice.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "order_items")
 public class OrderItem {
@@ -14,7 +16,7 @@ public class OrderItem {
 
     private Integer quantity;
 
-    private Double priceAtPurchase;
+    private BigDecimal priceAtPurchase;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "order_id")
@@ -55,11 +57,11 @@ public class OrderItem {
         this.quantity = quantity;
     }
 
-    public Double getPriceAtPurchase() {
+    public BigDecimal getPriceAtPurchase() {
         return priceAtPurchase;
     }
 
-    public void setPriceAtPurchase(Double priceAtPurchase) {
+    public void setPriceAtPurchase(BigDecimal priceAtPurchase) {
         this.priceAtPurchase = priceAtPurchase;
     }
 }
